Added -WX compiler flag for Windows build (same as -Werror: treat warnings as error

+++ libmariadb/ma_string.c
@@ -96,7 +96,7 @@ my_bool ma_dynstr_append_quoted(DYNAMIC_STRING *str,
                              const char *append, size_t len,
                              char quote)
 {
-  uint additional= str->alloc_increment;
+  uint additional= (uint)str->alloc_increment;
   uint lim= additional;
   uint i;
